Layout now based on bootstrap
Reworked the layout to look much nicer. There is now a navbar, and the
game widget is centered and shiny. The Bootstrap game widget doesn't do
anything yet, and we will need to write some AJAX in order of it to do
so.

At the moment, the game page looks funky because it has the old form
widget as well as the shiny Bootstrap one.

One subtle change is that in order to set up the navbar, we wrap in a
.container div in the default-layout rather than the
default-layout-wrapper.